+++ src/main/java/com/hotels/beans/transformer/TransformerImpl.java
@@ -23,170 +23,40 @@ import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.IntStream.range;
 
-import static javax.validation.Validation.buildDefaultValidatorFactory;
-
-import static org.apache.commons.lang3.StringUtils.SPACE;
 import static org.apache.commons.lang3.StringUtils.isNotEmpty;
 
 import static com.hotels.beans.constant.Punctuation.DOT;
-import static com.hotels.beans.constant.Punctuation.SEMICOLON;
 import static com.hotels.beans.constant.Punctuation.COMMA;
 import static com.hotels.beans.constant.Punctuation.LPAREN;
 import static com.hotels.beans.constant.Punctuation.RPAREN;
 import static com.hotels.beans.constant.ClassType.MIXED;
 import static com.hotels.beans.constant.ClassType.MUTABLE;
-import static com.hotels.beans.utils.ValidationUtils.notNull;
 import static com.hotels.beans.base.Defaults.defaultValue;
-import static com.hotels.beans.cache.CacheManagerFactory.getCacheManager;
 import static com.hotels.beans.populator.PopulatorFactory.getPopulator;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Parameter;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.function.Function;
-
-import javax.validation.ConstraintViolation;
-import javax.validation.Validator;
 
 import com.hotels.beans.annotation.ConstructorArg;
-import com.hotels.beans.cache.CacheManager;
 import com.hotels.beans.constant.ClassType;
 import com.hotels.beans.error.InvalidBeanException;
 import com.hotels.beans.error.MissingFieldException;
-import com.hotels.beans.model.FieldMapping;
-import com.hotels.beans.model.FieldTransformer;
-import com.hotels.beans.utils.ClassUtils;
-import com.hotels.beans.utils.ReflectionUtils;
 
 /**
  * Utility methods for populating Mutable, Immutable and Hybrid JavaBeans properties via reflection.
  * The implementations are provided by BeanUtils.
  */
-public class TransformerImpl implements Transformer {
-    /**
-     * Reflection utils class {@link ReflectionUtils}.
-     */
-    private final ReflectionUtils reflectionUtils;
-
-    /**
-     * Class reflection utils class {@link ClassUtils}.
-     */
-    private final ClassUtils classUtils;
-
-    /**
-     * CacheManager class {@link CacheManager}.
-     */
-    private final CacheManager cacheManager;
-
-    /**
-     * Contains both the field name mapping and the lambda function to be applied on fields.
-     */
-    private final TransformerSettings transformerSettings;
-
-    /**
-     * Default constructor.
-     */
-    public TransformerImpl() {
-        this.reflectionUtils = new ReflectionUtils();
-        this.classUtils = new ClassUtils();
-        this.transformerSettings = new TransformerSettings();
-        this.cacheManager = getCacheManager("transformer");
-    }
-
+public class TransformerImpl extends AbstractTransformer {
     /**
      * {@inheritDoc}
      */
     @Override
-    public final Transformer withFieldMapping(final FieldMapping... fieldMapping) {
-        final Map<String, String> fieldsNameMapping = transformerSettings.getFieldsNameMapping();
-        for (FieldMapping mapping : fieldMapping) {
-            fieldsNameMapping.put(mapping.getDestFieldName(), mapping.getSourceFieldName());
-        }
-        return this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void removeFieldMapping(final String destFieldName) {
-        notNull(destFieldName, "The field name for which the mapping has to be removed cannot be null!");
-        transformerSettings.getFieldsNameMapping().remove(destFieldName);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void resetFieldsMapping() {
-        transformerSettings.getFieldsNameMapping().clear();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @SuppressWarnings("unchecked")
-    public final Transformer withFieldTransformer(final FieldTransformer... fieldTransformer) {
-        Map<String, Function<Object, Object>> fieldsTransformers = transformerSettings.getFieldsTransformers();
-        for (FieldTransformer transformer : fieldTransformer) {
-            fieldsTransformers.put(transformer.getDestFieldName(), transformer.getTransformerFunction());
-        }
-        return this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void removeFieldTransformer(final String destFieldName) {
-        notNull(destFieldName, "The field name for which the transformer function has to be removed cannot be null!");
-        transformerSettings.getFieldsTransformers().remove(destFieldName);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void resetFieldsTransformer() {
-        transformerSettings.getFieldsTransformers().clear();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final Transformer setDefaultValueForMissingField(final boolean useDefaultValue) {
-        transformerSettings.setSetDefaultValue(useDefaultValue);
-        return this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final <T, K> K transform(final T sourceObj, final Class<? extends K> targetClass) {
-        notNull(sourceObj, "The object to copy cannot be null!");
-        notNull(targetClass, "The destination class cannot be null!");
-        return transform(sourceObj, targetClass, null);
-    }
-
-    /**
-     * Copies all properties from an object to a new one.
-     * @param sourceObj the source object
-     * @param targetClass the destination object class
-     * @param breadcrumb the full path of the current field starting from his ancestor
-     * @param <T> the Source object type
-     * @param <K> the target object type
-     * @return a copy of the source object into the destination object
-     */
-    private <T, K> K transform(final T sourceObj, final Class<? extends K> targetClass, final String breadcrumb) {
+    protected final <T, K> K transform(final T sourceObj, final Class<? extends K> targetClass, final String breadcrumb) {
         final K k;
-        final ClassType classType = classUtils.getClassType(targetClass);
+        final ClassType classType = getClassUtils().getClassType(targetClass);
         if (classType.is(MUTABLE)) {
             try {
                 k = targetClass.getDeclaredConstructor().newInstance();
@@ -197,12 +67,12 @@ public class TransformerImpl implements Transformer {
                 throw new InvalidBeanException(e.getMessage(), e);
             }
         } else {
-            k = injectValues(sourceObj, targetClass, classUtils.getAllArgsConstructor(targetClass), breadcrumb);
+            k = injectValues(sourceObj, targetClass, getClassUtils().getAllArgsConstructor(targetClass), breadcrumb);
             if (classType.is(MIXED)) {
                 injectNotFinalFields(sourceObj, k, breadcrumb);
             }
         }
-        validate(k);
+        getValidationUtils().validate(k);
         return k;
     }
 
@@ -257,10 +127,10 @@ public class TransformerImpl implements Transformer {
      */
     private <K> boolean canBeInjectedByConstructorParams(final Constructor constructor, final Class<K> targetClass) {
         final String cacheKey = "CanBeInjectedByConstructorParams-" + constructor.getDeclaringClass().getCanonicalName();
-        return ofNullable(cacheManager.getFromCache(cacheKey, Boolean.class)).orElseGet(() -> {
-            final boolean res = classUtils.getPrivateFinalFields(targetClass).size() == constructor.getParameterCount()
-                    && (classUtils.areParameterNamesAvailable(constructor) || classUtils.allParameterAnnotatedWith(constructor, ConstructorArg.class));
-            cacheManager.cacheObject(cacheKey, res);
+        return ofNullable(getCacheManager().getFromCache(cacheKey, Boolean.class)).orElseGet(() -> {
+            final boolean res = getClassUtils().getPrivateFinalFields(targetClass).size() == constructor.getParameterCount()
+                    && (getClassUtils().areParameterNamesAvailable(constructor) || getClassUtils().allParameterAnnotatedWith(constructor, ConstructorArg.class));
+            getCacheManager().cacheObject(cacheKey, res);
             return res;
         });
     }
@@ -278,19 +148,19 @@ public class TransformerImpl implements Transformer {
      * @throws InvalidBeanException {@link InvalidBeanException} if there is an error while retrieving the constructor args parameter
      */
     private <T, K> Object[] getConstructorArgsValues(final T sourceObj, final Class<K> targetClass, final Constructor constructor, final String breadcrumb) {
-        final Parameter[] constructorParameters = classUtils.getConstructorParameters(constructor);
+        final Parameter[] constructorParameters = getClassUtils().getConstructorParameters(constructor);
         final Object[] constructorArgsValues = new Object[constructorParameters.length];
         range(0, constructorParameters.length)
                 //.parallel()
                 .forEach(i -> {
                     String destFieldName = getDestFieldName(constructorParameters[i], targetClass.getCanonicalName());
                     if (isNull(destFieldName)) {
-                        constructorArgsValues[i] =  classUtils.getDefaultTypeValue(constructorParameters[i].getType());
+                        constructorArgsValues[i] =  getClassUtils().getDefaultTypeValue(constructorParameters[i].getType());
                     } else {
                         String sourceFieldName = getSourceFieldName(destFieldName);
                         constructorArgsValues[i] =
-                                ofNullable(getFieldValue(sourceObj, sourceFieldName, targetClass, classUtils.getDeclaredField(targetClass, destFieldName), breadcrumb))
-                                .orElse(classUtils.getDefaultTypeValue(constructorParameters[i].getType()));
+                                ofNullable(getFieldValue(sourceObj, sourceFieldName, targetClass, getClassUtils().getDeclaredField(targetClass, destFieldName), breadcrumb))
+                                .orElse(getClassUtils().getDefaultTypeValue(constructorParameters[i].getType()));
                     }
                 });
         return constructorArgsValues;
@@ -311,7 +181,7 @@ public class TransformerImpl implements Transformer {
      * @return the source field name.
      */
     private String getSourceFieldName(final String fieldName) {
-        return ofNullable(transformerSettings.getFieldsNameMapping().get(fieldName)).orElse(fieldName);
+        return ofNullable(getTransformerSettings().getFieldsNameMapping().get(fieldName)).orElse(fieldName);
     }
 
     /**
@@ -322,17 +192,17 @@ public class TransformerImpl implements Transformer {
      */
     private String getDestFieldName(final Parameter constructorParameter, final String declaringClassName) {
         String cacheKey = "DestFieldName-" + declaringClassName + "-" + constructorParameter.getName();
-        return ofNullable(cacheManager.getFromCache(cacheKey, String.class))
+        return ofNullable(getCacheManager().getFromCache(cacheKey, String.class))
                 .orElseGet(() -> {
                     String destFieldName;
                     if (constructorParameter.isNamePresent()) {
                         destFieldName = constructorParameter.getName();
                     } else {
-                        destFieldName = ofNullable(reflectionUtils.getParameterAnnotation(constructorParameter, ConstructorArg.class, declaringClassName))
+                        destFieldName = ofNullable(getReflectionUtils().getParameterAnnotation(constructorParameter, ConstructorArg.class, declaringClassName))
                                 .map(ConstructorArg::value)
                                 .orElse(null);
                     }
-                    cacheManager.cacheObject(cacheKey, destFieldName);
+                    getCacheManager().cacheObject(cacheKey, destFieldName);
                     return destFieldName;
                 });
     }
@@ -349,7 +219,7 @@ public class TransformerImpl implements Transformer {
      * @throws InvalidBeanException {@link InvalidBeanException} if an error occurs while retrieving the value
      */
     private <T, K> Object[] getConstructorValuesFromFields(final T sourceObj, final Class<K> targetClass, final String breadcrumb) {
-        final List<Field> declaredFields = classUtils.getDeclaredFields(targetClass, true);
+        final List<Field> declaredFields = getClassUtils().getDeclaredFields(targetClass, true);
         return declaredFields.stream()
                 .map(field -> getFieldValue(sourceObj, targetClass, field, breadcrumb))
                 .toArray(Object[]::new);
@@ -367,9 +237,9 @@ public class TransformerImpl implements Transformer {
      */
     private <T, K> void injectNotFinalFields(final T sourceObj, final K targetObject, final String breadcrumb) {
         final Class<?> targetObjectClass = targetObject.getClass();
-        classUtils.getNotFinalFields(targetObjectClass, true)
+        getClassUtils().getNotFinalFields(targetObjectClass, true)
                 //.parallelStream()
-                .forEach(field -> reflectionUtils.setFieldValue(targetObject, field, getFieldValue(sourceObj, targetObjectClass, field, breadcrumb)));
+                .forEach(field -> getReflectionUtils().setFieldValue(targetObject, field, getFieldValue(sourceObj, targetObjectClass, field, breadcrumb)));
     }
 
     /**
@@ -404,13 +274,13 @@ public class TransformerImpl implements Transformer {
         Object fieldValue = null;
         String fieldBreadcrumb = evalBreadcrumb(field.getName(), breadcrumb);
         if (isNotEmpty(sourceFieldName)) {
-            boolean primitiveType = classUtils.isPrimitiveType(field.getType());
-            boolean isFieldTransformerDefined = transformerSettings.getFieldsTransformers().containsKey(field.getName());
+            boolean primitiveType = getClassUtils().isPrimitiveType(field.getType());
+            boolean isFieldTransformerDefined = getTransformerSettings().getFieldsTransformers().containsKey(field.getName());
             fieldValue = getSourceFieldValue(sourceObj, sourceFieldName, field, isFieldTransformerDefined);
             if (nonNull(fieldValue)) {
                 // is not a primitive type or an optional && there are no transformer function
                 // defined it recursively evaluate the value
-                boolean notPrimitiveAndNotSpecialType = !primitiveType && !classUtils.isSpecialType(field.getType());
+                boolean notPrimitiveAndNotSpecialType = !primitiveType && !getClassUtils().isSpecialType(field.getType());
                 if ((notPrimitiveAndNotSpecialType || Optional.class.isAssignableFrom(fieldValue.getClass()))
                         && !isFieldTransformerDefined) {
                     fieldValue = getFieldValue(targetClass, field, fieldValue, fieldBreadcrumb);
@@ -444,9 +314,9 @@ public class TransformerImpl implements Transformer {
     private <T> Object getSourceFieldValue(final T sourceObj, final String sourceFieldName, final Field field, final boolean isFieldTransformerDefined) {
         Object fieldValue = null;
         try {
-            fieldValue = reflectionUtils.getFieldValue(sourceObj, sourceFieldName, field.getType());
+            fieldValue = getReflectionUtils().getFieldValue(sourceObj, sourceFieldName, field.getType());
         } catch (MissingFieldException e) {
-            if (!isFieldTransformerDefined && !transformerSettings.isSetDefaultValue()) {
+            if (!isFieldTransformerDefined && !getTransformerSettings().isSetDefaultValue()) {
                 throw e;
             }
         } catch (Exception e) {
@@ -465,8 +335,8 @@ public class TransformerImpl implements Transformer {
      * @return the transformed field.
      */
     private Object getTransformedField(final Field field, final String breadcrumb, final Object fieldValue) {
-        String fieldName = transformerSettings.isFlatFieldNameTransformation() ? field.getName() : breadcrumb;
-        return ofNullable(transformerSettings.getFieldsTransformers().get(fieldName))
+        String fieldName = getTransformerSettings().isFlatFieldNameTransformation() ? field.getName() : breadcrumb;
+        return ofNullable(getTransformerSettings().getFieldsTransformers().get(fieldName))
                 .map(fieldTransformer -> fieldTransformer.apply(fieldValue))
                 .orElse(fieldValue);
     }
@@ -489,39 +359,5 @@ public class TransformerImpl implements Transformer {
                         // recursively inject object
                         transform(fieldValue, field.getType(), breadcrumb)
                 );
-    }
-
-    /**
-     * Checks if an object is valid.
-     * @param k the object to check
-     * @param <K> the object class
-     * @throws InvalidBeanException {@link InvalidBeanException} if the validation fails
-     */
-    private <K> void validate(final K k) {
-        final Set<ConstraintViolation<Object>> constraintViolations = getValidator().validate(k);
-        if (!constraintViolations.isEmpty()) {
-            final String errors = constraintViolations.stream()
-                    .map(cv -> cv.getRootBeanClass().getCanonicalName()
-                            + DOT.getSymbol()
-                            + cv.getPropertyPath()
-                            + SPACE
-                            + cv.getMessage())
-                    .collect(joining(SEMICOLON.getSymbol()));
-            throw new InvalidBeanException(errors);
-        }
-    }
-
-    /**
-     * Creates the validator.
-     * @return a {@link Validator} instance.
-     */
-    private Validator getValidator() {
-        String cacheKey = "BeanValidator";
-        return ofNullable(cacheManager.getFromCache(cacheKey, Validator.class))
-                .orElseGet(() -> {
-                    Validator validator = buildDefaultValidatorFactory().getValidator();
-                    cacheManager.cacheObject(cacheKey, validator);
-                    return validator;
-                });
     }
 }